requirements.txt
astroid==2.3.3
beautifulsoup4==4.8.2
blessings==1.7
certifi==2019.11.28
chardet==3.0.4
codecov==2.0.15
coloredlogs==10.0
colour-runner==0.1.1
coverage==5.0.3
deepdiff==4.0.9
demjson==2.2.4
diskcache==4.1.0
filelock==3.0.12
humanfriendly==4.18
idna==2.8
importlib-metadata==1.4.0
isort==4.3.21
jsonpickle==1.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.1.0
ordered-set==3.1.1
packaging==20.0
pluggy==0.13.1
py==1.8.1
Pygments==2.5.2
pylint==2.4.4
pyparsing==2.4.6
PySnooper==0.3.0
pyTelegramBotAPI==3.6.6
python-dotenv==0.10.5
requests==2.22.0
rootpath==0.1.1
six==1.13.0
soupsieve==1.9.5
termcolor==1.1.0
toml==0.10.0
tox==3.14.3
typed-ast==1.4.0
urllib3==1.25.7
virtualenv==16.7.9
wrapt==1.11.2
zipp==2.0.0
